# implementation/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenants/{tenantId}/attestation-providers": {
      "post": {
        "operationId": "IssuanceConfigurationController_createAttestationProvider",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register an Organisation as an Attestation Provider (TEST)",
        "tags": [
          "issuance-configuration"
        ]
      },
      "get": {
        "operationId": "TenantListingController_attestationProviders",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List Attestation Providers",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/attestation-providers/{providerId}/provision": {
      "delete": {
        "operationId": "IssuanceConfigurationController_decommission",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Release the provider's engine tenant so another provider can be given it",
        "tags": [
          "issuance-configuration"
        ]
      },
      "post": {
        "operationId": "IssuanceConfigurationController_provision",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Provision the engine tenant, signing key and registration certificate",
        "tags": [
          "issuance-configuration"
        ]
      }
    },
    "/v1/tenants/{tenantId}/attestation-providers/{providerId}/provider-authentication": {
      "get": {
        "operationId": "IssuanceConfigurationController_providerAuthentication",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "What a Wallet would see when authenticating this provider (ARF §6.6.2.2)",
        "tags": [
          "issuance-configuration"
        ]
      }
    },
    "/v1/tenants/{tenantId}/credential-types": {
      "post": {
        "operationId": "IssuanceConfigurationController_createCredentialType",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Define a credential type and the Rulebook that governs it",
        "tags": [
          "issuance-configuration"
        ]
      },
      "get": {
        "operationId": "TenantListingController_credentialTypes",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List credential types",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/credential-types/{credentialTypeId}": {
      "get": {
        "operationId": "IssuanceConfigurationController_getCredentialType",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credentialTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read a credential type",
        "tags": [
          "issuance-configuration"
        ]
      }
    },
    "/v1/tenants/{tenantId}/issuance-policies": {
      "post": {
        "operationId": "IssuanceConfigurationController_createPolicy",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create an issuance policy",
        "tags": [
          "issuance-configuration"
        ]
      },
      "get": {
        "operationId": "TenantListingController_issuancePolicies",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List issuance policies",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/tenants/{tenantId}/issuance-policies/{policyId}": {
      "get": {
        "operationId": "IssuanceConfigurationController_getPolicy",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read an issuance policy and its versions",
        "tags": [
          "issuance-configuration"
        ]
      }
    },
    "/v1/tenants/{tenantId}/issuance-policies/{policyId}/versions": {
      "post": {
        "operationId": "IssuanceConfigurationController_createVersion",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create, and optionally publish, an issuance policy version",
        "tags": [
          "issuance-configuration"
        ]
      }
    },
    "/v1/issuances": {
      "post": {
        "operationId": "IssuanceController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create an issuance and obtain a credential offer",
        "tags": [
          "issuances"
        ]
      },
      "get": {
        "operationId": "TransactionListingController_issuances",
        "parameters": [
          {
            "name": "policyId",
            "required": false,
            "in": "query",
            "description": "Narrows to one issuance policy.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List issuance transactions",
        "tags": [
          "Listing"
        ]
      }
    },
    "/v1/issuances/{issuanceId}": {
      "get": {
        "operationId": "IssuanceController_get",
        "parameters": [
          {
            "name": "issuanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read an issuance transaction",
        "tags": [
          "issuances"
        ]
      }
    },
    "/v1/issued-credentials/{issuedCredentialId}/revoke": {
      "post": {
        "operationId": "IssuanceController_revoke",
        "parameters": [
          {
            "name": "issuedCredentialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Revoke an issued attestation. Irreversible — `AS-AP-07-007` (`VCR_04`)",
        "tags": [
          "issuances"
        ]
      }
    },
    "/v1/issued-credentials/{issuedCredentialId}/status": {
      "post": {
        "operationId": "IssuanceController_changeStatus",
        "parameters": [
          {
            "name": "issuedCredentialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Suspend or reinstate an issued attestation. `VALID` only from `SUSPENDED`",
        "tags": [
          "issuances"
        ]
      }
    },
...
  },
  "info": {
    "title": "European Digital Trust Platform — Verification as a Service",
    "description": "V0 business API. Customers work with presentation policies and transactions; protocol details (DCQL, OpenID4VP, engine sessions) are not part of this contract. No ARF or Technical Specification conformance and no production readiness is claimed.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tenant API key"
      }
    },
...
  }
}